Fixed requirements.txt Moved validate_path() to osutil.py Updated file_manager.py
